+++ packages/package_info_plus/CHANGELOG.md
@@ -1,9 +1,10 @@
-## NEXT
 ## 1.0.3
 
 * Update package_info_plus to 4.0.1.
 * Update package_info_plus_platform_interface to 2.0.1.
 * Increase the minimum Flutter version to 3.3.
+* Update the example app and integration_test.
+* Minor cleanups.
 
 ## 1.0.2